WIP: Accommodate the final cursor position for compatibility
The cursor should follow the effects of "H" or "^" commands
for its final placement; unfortunately, wrapped lines can be
scrolled with "H" (or "L") for a non-zero &scrolloff value
(see selftestdir/dots_04); additionally, folded lines that
are written at the top require another cursor adjustment and
the "[count]jH" movement can sometimes contribute to a wrong
approximation of "H" later with "Ctrl-{e,y}".

Therefore, confine "H" to folded lines when &so is non-zero
and otherwise rely on "Ctrl-{e,y}" for the effect of "H".

(TODO: This commit is for convenience only, parts of it
belong to former HEAD and HEAD~2.)